Set up the PostgreSQL container with a password
Because:
- Apparently, newer Postgres images from the DockerHub require a
  POSTGRES_PASSWORD environment variable to start up from scratch.
- ecadd7781e0d648905dc70df0504ec3dd0494dc8 changed the Postgres image
  version to 12, thereby introducing this problem.

Older versions of the postgres image would use trust authentication
within the local docker network. Not so anymore, unless specifically
specified. This is still no safer than the old way of doing it, since we
are using passwords in the clear here. However, it remains safe enough
as long as only the containers in this Docker Compose project can access
the postgres container.